+++ mycli/packages/special/utils.py
@@ -155,6 +155,7 @@ def format_connection_dsn(
     database: str | None,
     socket: str | None,
     character_set: str | None,
+    boundary_id: str | None = None,
     ssh_jump: str | None = None,
     vault_address: str | None = None,
     vault_mount: str | None = None,
@@ -174,6 +175,8 @@ def format_connection_dsn(
         port_part = ''
     if character_set and character_set != 'utf8mb4':
         query_part['character_set'] = character_set
+    if boundary_id:
+        query_part['boundary_id'] = boundary_id
     if ssh_jump:
         query_part['ssh_jump'] = ssh_jump
     if vault_address: